Initial scripts added
Script to change the resolution of a raster to decrease elevation variation in flights. Script to buffer out from objects to make the drone fly up sooner when it runs into an object. Script to smooth pixels to make flight better. Script to run these three scripts in order. Script to calculate drone flight plan outline for multi-angle flights for each flight.